+++ store/src/main/java/discord4j/store/primitive/ForwardingStoreService.java
@@ -51,7 +51,7 @@ public class ForwardingStoreService implements StoreService {
     }
 
     @Override
-    public short priority() {
+    public int priority() {
         return toForward.priority();
     }
